Refactor code associated with Author authentication into a concern
